% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqvalue.R
\name{as.data.frame.fqvalue}
\alias{as.data.frame.fqvalue}
\title{Convert an fqvalue object to a data frame}
\usage{
\method{as.data.frame}{fqvalue}(x, ...)
}
\arguments{
\item{x}{"fqvalue" object}

\item{...}{extra arguments, not used}
}
\description{
Convert an fqvalue object to a data frame
}
