% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fqvalue.R
\name{as.double.fqvalue}
\alias{as.double.fqvalue}
\title{Convert an fqvalue object to a vector of fq-values}
\usage{
\method{as.double}{fqvalue}(x, ...)
}
\arguments{
\item{x}{"fqvalue" object}

\item{...}{extra arguments, not used}
}
\description{
Convert an fqvalue object to a vector of fq-values
}

