% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fqvalue.R
\name{compare_qvalue}
\alias{compare_qvalue}
\title{Compare functional q-value to traditional q-values}
\usage{
compare_qvalue(fq, ...)
}
\arguments{
\item{fq}{An fqvalue object}

\item{...}{additional arguments passed to the traditional qvalue function}
}
\value{
ggplot object (print to show the plot)
}
\description{
Compare traditional qvalue to functional q-values in a scatterplot,
colored either by the quantile of the informative variable or by an oracle procedure
}
\details{
If the column \code{oracle} exists in the fqvalue, \code{compare_qvalue}
will use that as the colors of the points. Otherwise, the color depends
on the quantiles of the informative variable.
}
