% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_factorial}
\alias{do_factorial}
\title{Call a function with all factorial combinations of given parameters}
\usage{
do_factorial(.data, expr, ...)
}
\arguments{
\item{.data}{Data within which the expression is evaluated}

\item{expr}{Expression to be evaluated, which must be a single
function call}

\item{...}{One or more vectors of parameters that will be given
to the function in combination}
}
\value{
A data.frame, grouped by the given factorial parameters
}
\description{
An extension of dplyr's philosophy. Essential in factorial
simulations and applications
}

