% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_fixed_pi0.R
\name{estimate_fixed_pi0}
\alias{estimate_fixed_pi0}
\title{Estimate a constant functional proportion}
\usage{
estimate_fixed_pi0(p, z0, lambda = seq(0, 0.9, 0.05), tau = seq(0, 0.9,
  0.05), pi0.method = "smoother", lambda.df = 3, tau.df = 3)
}
\arguments{
\item{p}{A vector of p-values}

\item{z0}{A vector of observations from the informative variable, of the same length as \code{p}}

\item{lambda}{Choices of the tuning parameter "lambda" (for the p-value) used to estimate pi0(z)}

\item{tau}{Choices of the tuning parameter "tau" (for the informative variable) used to estiamte pi0(z)}

\item{pi0.method}{The method used to estimate pi0(z), being either "smoother" (default) or "bootstrap". The method depends on the tuning parameters "lambda" and "tau"}

\item{lambda.df}{Degrees of freedom for smoother in argument "lambda"}

\item{tau.df}{Degrees of freedom for smoother in argument "tau"}
}
\value{
Estimate of pi0(z)
}
\description{
Estimate the functional proportion pi0(z) when it is independent of z (whcih is the quantile transform variable of the informative variable z0). For information on z, please refer to \code{fqvalue} or \code{estimate_fpi0}
}
