% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_fixed_pi0.R
\name{fixed_pi0_table}
\alias{fixed_pi0_table}
\title{Provide estimates of the functional proportion at varying values of lambda and tau}
\usage{
fixed_pi0_table(p, z0, lambda = seq(0, 0.9, 0.05), tau = seq(0, 0.9,
  0.05), pi0.method = NULL, lambda.df = 3, tau.df = 3)
}
\arguments{
\item{p}{A vector of p-values}

\item{z0}{A vector of observations from the informative variable, of the same length as \code{p}}

\item{lambda}{Choices of the tuning parameter "lambda" (for the p-value) used to estimate pi0(z)}

\item{tau}{Choices of the tuning parameter "tau" (for the informative variable) used to estiamte pi0(z)}

\item{pi0.method}{The method used to estimate pi0(z), being either "smoother" (default) or "bootstrap". The method depends on the tuning parameters "lambda" and "tau"}

\item{lambda.df}{Degrees of freedom for smoother in argument "lambda"}

\item{tau.df}{Degrees of freedom for smoother in argument "tau"}
}
\value{
A data.frame with the following columns

\item{lambda}{Choices of lambda as the p-value threshold}
\item{tau}{Choices of tau as the z threshold, where z is the quantile transformed z0}
\item{L}{The number of hypotheses for which p > lambda and z > tau}
\item{pi0hat}{The estimate of pi0(z) using these choices of lambda and tau}
\item{variance}{The estimated variance of the estimate of pi0(z)}
\item{se}{The estimated standard error of the estimate of pi0(z)}
\item{bias}{The estimated bias of the estimate of pi0(z)}
\item{MSEhat}{The estimated mean squared error (MSE) of the estimate of pi0(z)}
}
\description{
Estimate the functional proportion pi0(z) at varying values of lambda and tau when pi0(z) is independent of z
}
