% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{fqvalue-class}
\alias{fqvalue-class}
\title{Functional q-values calculated for a set of p-values and z}
\description{
This controls false discovery rate taking into consideration the
dependence of p-values on a factor z.
}
\section{Slots}{

 \describe{
   \item{\code{table}:}{a \link{data.table} containing the p-values, z, z0, estimated fpi0, estimated density, local FDR, and the computed f-qvalues.}
   \item{\code{fPi0}:}{An \linkS4class{fPi0} object, or NULL if functional pi0 estimation was not used}
   \item{\code{density}:}{A \link{data.table} containing the estimated density of each point in the p-value/z space, used to calculate the local FDR}
}
}

