% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqvalue.R
\name{fqvalue}
\alias{fqvalue}
\title{Estimate functional q-value based on p-value and an informative variable}
\usage{
fqvalue(p.value, z0, pi0.method = "gam", lambda = seq(0.4, 0.9, 0.1),
  fixed.pi0 = FALSE, monotone.window = 0.01, ...)
}
\arguments{
\item{p.value}{A vector of p-values}

\item{z0}{A vector of observations from the informative variable, of the same length as \code{p}}

\item{pi0.method}{Method for estimating the functional proportion pi0(z); either "gam" (default), "glm", "kernel" or "bin"}

\item{lambda}{The parameter Lambda or its values to use in estimating pi0(z)}

\item{fixed.pi0}{Whether pi0(z) is believed to be independent of z}

\item{monotone.window}{Parameter used to force estimated densities to
decrease with increasing p-values; higher value of this parameter means densities are smoothed
more aggressively. If NULL, perform no such smoothing}

\item{...}{Extra arguments to be passed to kernelUnitInterval for estimating the density}
}
\value{
An object of S3 class "fqvalue"
}
\description{
Estimate functional q-values based on a vector of p-values and a vector of realizations of
the informative variable z0, where z0 is known to be related to either the power of a statistical test or the likelihood of a true null hypothesis. 
Assume there are m observations z0_i, i=1,...,m of the informative variable z0. They are quantile transformed into z_i, i=1,...,m, such that z_i = rank(z0_i) / m, where rank(z0_i) is the rank of z0_i among z0_i, i=1,...,m. Doing so ensures that z_i, i=1,...,m are approximately uniform distributed on the interval [0,1], Correspondingly, z0 has been quantile transformed into z, such that z is approximately uniformly distributed on [0,1].
The functional proportion is regarded as a function of z and is denoted by pi0(z), and the fFDR methodology is applied to the p-value p and z.
}
