% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fpi0.R
\name{plot.fPi0}
\alias{plot.fPi0}
\title{Plot the estimated functional proportion}
\usage{
\method{plot}{fPi0}(x, horizontal.line = FALSE, subsample = 10000, ...)
}
\arguments{
\item{x}{fPi0 object}

\item{horizontal.line}{Whether to draw a horizontal line showing the
Storey estimate of a constant pi0(z) for each choice of lambda}

\item{subsample}{Maximum number of points to plot; otherwise randomly sampled
(if too many points are plotted, it is impossible to distinguish dotted
and solid lines)}

\item{...}{Other plotting parameters (not used)}
}
\value{
A ggplot2 graph plotting pi0(z) as a function of z. For information on z, please refer to \code{fqvalue} or \code{estimate_fpi0}
}
\description{
Plot the estimated functional proportion pi0(z) for different choices of tuning parameter "lambda", with reference as Storey's estimate of a constant pi0(z).
}
