\docType{methods}
\name{plotMISE,fqvalue-method}
\alias{plotMISE,fqvalue-method}
\title{Plot demonstrating how lambda is chosen to minimize MISE in
estimation of pi0(z)}
\usage{
\S4method{plotMISE}{fqvalue}(x, ...)
}
\arguments{
  \item{x}{fqvalue object}
}
\description{
Performs a plotMISE plot of the functional pi0 estimate
contained within an fqvalue object
}

