% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_fpi0.R
\name{predict.fPi0}
\alias{predict.fPi0}
\title{Predict the functional proportion}
\usage{
\method{predict}{fPi0}(object, z0 = NULL, z = NULL, lambda = NULL, ...)
}
\arguments{
\item{object}{fPi0 object}

\item{z0}{New z0 values (when they are between 0 and 1)}

\item{z}{New z values (as the quantile transformed values of z0)}

\item{lambda}{The value of the tuning parameter to be used for prediction. If null, defaults to the lambda
chosen in the fPi0 object}

\item{...}{Extra arguments, not used}
}
\value{
Vector of fPi0 predictions
}
\description{
Predict the values of the functional proportion when it is evaluated at a vector z' whose entries are between 0 and 1.
}
\details{
Assume the random variable z0 may affect the power of a statistical test (that induces the p-values) or the likelihood of a true null hypothesis. The m observations z0_i, i=1,...,m of z0 are quantile transformed into z_i, i=1,...,m such that z_i = rank(z0_i) / m, where rank(z0_i) is the rank of z0_i among z0_i, i=1,...,m. Consequently, z_i, i=1,...,m are approximately uniformly distributed on the interval [0,1]. When z_i, i=1,...,m are regarded as observations from the random variable z, then z is approximately uniformly distributed on [0,1]. Namely, z0 has been quantile transformed into z, and they are equivalent. Further, z or z0 is referred to as the informative variable.
}
