% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate_fpi0.R
\name{print.fPi0}
\alias{print.fPi0}
\title{Extract functional pi0 estimates.}
\usage{
\method{print}{fPi0}(x, ...)
}
\arguments{
\item{x}{"fPi0" object}

\item{...}{Extra arguments, not used}
}
\description{
Extract functional pi0 estimates.
}

