% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqvalue.R
\name{print.fqvalue}
\alias{print.fqvalue}
\title{Print an fqvalue object}
\usage{
\method{print}{fqvalue}(x, ...)
}
\arguments{
\item{x}{"fqvalue" object}

\item{...}{Extra arguments, not used}
}
\description{
Print an fqvalue object
}
