% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reverseloglog_trans}
\alias{reverseloglog_trans}
\title{log(-log(x)) transformation for a ggplot2 scale}
\usage{
reverseloglog_trans(base = exp(1))
}
\arguments{
\item{base}{Base used in the exponentiation}
}
\description{
this code is a modified version of this:
http://stackoverflow.com/questions/11053899
}

