% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_t_tests}
\alias{simulate_t_tests}
\title{simulate one-sample t-tests}
\usage{
simulate_t_tests(
  m = 4000,
  pi0 = 0.5,
  mu.sd = 0.3,
  mu.min = 0,
  n.lmean = 2,
  n.lsd = 2.5,
  sample.sd = 1,
  seed = NULL,
  ...
)
}
\arguments{
\item{m}{Number of hypotheses to simulate}

\item{pi0}{Either a value for the fraction of true null hypotheses for which mu=0,
or a vector each of whose entries is the likelihood of a true null hypothesis}

\item{mu.sd}{Standard deviation used to generate the means of the Normal distributions under the
alternative hypothesis}

\item{mu.min}{The minumum value of abs(mu) for the means of the Normal distributions under the alternative hypothesis}

\item{n.lmean}{Mean of the sample sizes on a log scale}

\item{n.lsd}{Standard deviation of the sample sizes on a log scale}

\item{sample.sd}{Standard deviation for the Normal distributions}

\item{seed}{Optionally, random seed to set before simulating}

\item{...}{Extra arguments, such as replication (may not be used)}
}
\value{
A data frame, one row per hypothesis, with the columns

\item{p.value}{Two-sided t-test p-value}
\item{n}{Sample size of each sample}
\item{mu}{Mean of each sample}
\item{oracle}{TRUE if the alternative hypothesis holds, and FALSE for a true null hypothesis}
}
\description{
\code{simulateTTests} simulates a number of one-sample t-tests on
Normally distributed samples with varying mean mu0 and sample size.
}
\details{
Each sample and test is constructed according to the following
generative process.

The mean of a Normal distribution, mu, is either 0 under the null hypothesis, or
generated as (\code{mu.min} + abs(Norm(0, \code{mu.sd})) under the alternative
hypothesis. So, the means are always either positive or 0.
\code{mu.min} can be used to ensure a minimal effect size between the null hypothesis and the
weakest alternative hypothesis (since by default an alternative hypothesis can
be arbitrarily close to being a true null hypothesis).

The sample size of each sample, n, is generated from a log-normal distribution, independent of the means of the Normal distributions. Specifically, it is generated as
2 + round(LNorm(n.lmean, n.lsd)), such that the sample size is never less
than 2 (which would make a t-test impossible).

At this point, a random sample of size n is generated from the
Norm(mu, \code{sample.sd}) distribution. A one-sample two-sided t-test
is performed and the p-value obtained.
}
