% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Simulation-class}
\alias{Simulation-class}
\title{Factorial simulation of the fqvalue package}
\description{
The results of the fqvalue package, including power and FDR, on simulated
t-tests, where the simulation and fqvalue parameters are set up in
factorial combinations
}
\section{Slots}{

 \describe{
   \item{\code{table}:}{a \link{data.table} containing the p-values, z, z0, estimated fpi0, estimated density, local FDR, and the computed f-qvalues.}
   \item{\code{fPi0}:}{An \linkS4class{fPi0} object, or NULL if functional pi0 estimation was not used}
   \item{\code{density}:}{A \link{data.table} containing the estimated density of each point in the p-value/z space, used to calculate the local FDR}
}
}

