% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\docType{methods}
\name{summary,Simulation-method}
\alias{summary,Simulation-method}
\title{Summarize a factorial fqvalue simulation}
\usage{
\S4method{summary}{Simulation}(object, alpha = 0.05, ...)
}
\arguments{
\item{object}{an fqvalueSimulation object}

\item{Desired}{confidence level}
}
\value{
a data.table summarizing the simulation
}
\description{
Summarize a factorial fqvalue simulation
}

