\name{summary.fqvalue}
\alias{summary.fqvalue}
\title{Summarize results of an fqvalue analysis}
\usage{
\method{summary}{fqvalue}(fq)
}
\arguments{
  \item{fq}{An object of class fqvalue.}
}
\value{
A data.table with columns summarizing the performance of
fqvalue
}
\description{
Summarize results of an fqvalue analysis
}

