\name{summary.fqvalueSimulation}
\alias{summary.fqvalueSimulation}
\title{Summarize a factorial fqvalue simulation}
\usage{
summary.fqvalueSimulation(object, alpha = 0.05, ...)
}
\arguments{
  \item{object}{an fqvalueSimulation object}

  \item{Desired}{confidence level}
}
\description{
Summarize a factorial fqvalue simulation
}

