% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{summary.Simulation}
\alias{summary.Simulation}
\title{Summarize a factorial fqvalue simulation}
\usage{
\method{summary}{Simulation}(object, alpha = 0.05, ...)
}
\arguments{
\item{object}{an fqvalueSimulation object}

\item{alpha}{Desired confidence level}

\item{...}{Extra arguments (not used)}
}
\value{
a data.table summarizing the simulation
}
\description{
Summarize a factorial fqvalue simulation
}

