% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast_data.R
\name{yeast_summary}
\alias{yeast_summary}
\alias{exp.e}
\alias{exp.g}
\alias{exp.pos}
\alias{marker}
\alias{marker.pos}
\title{Yeast eQTL data}
\format{
A RData object.
}
\source{
\href{https://journals.plos.org/plosbiology/article?id=10.1371/journal.pbio.0060083}{PLoS Biology}
}
\usage{
data(yeast)

\method{exp}{e}()

\method{exp}{g}()

\method{exp}{pos}()

marker()

marker.pos()
}
\description{
Data from an eQTL experiment on baker's yeast, with gene expressions in each of the 109 genotyped strains under two conditions, glucose and ethanol.
}
\examples{
library(fFDR)
data(yeast)

yeast_summary()

}
\references{
Smith, Erin N and Kruglyak, Leonid. (2008). Gene-environment interaction in yeast gene expression. PLoS Biology 6(4), e83.
(\href{https://journals.plos.org/plosbiology/article?id=10.1371/journal.pbio.0060083}{PLoS Biology})
}
\keyword{datasets}
