% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate_fixed_pi0.R
\name{estimate_fixed_pi0}
\alias{estimate_fixed_pi0}
\title{compute pi0 where the truth is independent of Z}
\usage{
estimate_fixed_pi0(p, z0, lambda = seq(0, 0.9, 0.05), tau = seq(0, 0.9,
  0.05), pi0.method = "smoother", lambda.df = 3, tau.df = 3)
}
\arguments{
\item{p}{a vector of p-values}

\item{z0}{a vector of a power-informative factor, of the same length as
\code{p}}

\item{lambda}{Possible values of lambda to try}

\item{tau}{Possible values of tau to try}

\item{pi0.method}{Either "smoother" (default) or "bootstrap", the method for
choosing pi0 based on the tuning parameters "lambda" and "tau"}

\item{lambda.df}{Degrees of freedom for smoother in lambda direction}

\item{tau.df}{Degrees of freedom for smoother in tau direction}
}
\value{
estimated pi0
}
\description{
\code{fixedPi0} estimates pi0 for a set of p-values using a power-informative
factor z0. This is the assumption that although power increases with z0,
the truth of the hypotheses is independent of it.
}

