% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate_fixed_pi0.R
\name{fixed_pi0_table}
\alias{fixed_pi0_table}
\title{compute a table of pi0hat estimates at varying values of lambda and tau}
\usage{
fixed_pi0_table(p, z0, lambda = seq(0, 0.9, 0.05), tau = seq(0, 0.9, 0.05),
  pi0.method = NULL, lambda.df = 3, tau.df = 3)
}
\arguments{
\item{p}{a vector of p-values}

\item{z0}{a vector of a power-informative factor, of the same length as
\code{p}}

\item{lambda}{Possible values of lambda to try}

\item{tau}{Possible values of tau to try}

\item{pi0.method}{Either "smoother" (default) or "bootstrap", the method for
choosing pi0 based on the tuning parameters "lambda" and "tau"}

\item{lambda.df}{Degrees of freedom for smoother in lambda direction}

\item{tau.df}{Degrees of freedom for smoother in tau direction}
}
\value{
a data.frame with the following columns

\item{lambda}{Choices of lambda, the p-value threshold}
\item{tau}{Choices of tau, the z threshold}
\item{L}{The number of hypotheses for which p > lambda and z > tau}
\item{pi0hat}{The estimate of pi0 using these choices of lambda and tau}
\item{variance}{The estimated variance of the pi0 estimate}
\item{se}{The estimated standard error of the pi0 estimate}
\item{bias}{The estimated bias of the pi0 estimate}
\item{MSEhat}{The estimated mean squared error of the pi0 estimate, typically used to choose MSE}
}
\description{
\code{fixedPi0Table} estimates pi0 for a set of p-values using a power-informative
factor z0. This is the assumption that although power increases with z0,
the truth of the hypotheses is independent of it.
}

