\docType{methods}
\name{plot,fPi0-method}
\alias{plot,fPi0-method}
\title{Given an fPi0 object, plot the estimates of pi_0(z) with multiple
choices of lambda}
\usage{
\S4method{plot}{fPi0}(x, ...)
}
\arguments{
  \item{x}{fPi0 object}

  \item{...}{other plotting parameters (not used)}
}
\value{
a ggplot2 graph plotting pi0 as a function of z
}
\description{
Given an fPi0 object, plot the estimates of pi_0(z) with
multiple choices of lambda
}

