% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.fqvalue.R
\name{plot.fqvalue}
\alias{plot.fqvalue}
\title{Scatterplot comparing the p-value to the quantile of Z}
\usage{
\method{plot}{fqvalue}(x, pi0 = TRUE, threshold = c(0.005, 0.01, 0.05, 0.1),
  cloglog = FALSE, ...)
}
\arguments{
\item{x}{fqvalue object}

\item{pi0}{Should a second panel of the estimate of pi0(z) be shown
underneath the scatterplot}

\item{threshold}{vector of thresholds used for colors on the graph
(default c(.005, .01, .05, .1))}

\item{cloglog}{whether the plot should be on a log(-log(pvalue)) scale
(default FALSE)}

\item{...}{Extra arguments (not used)}
}
\value{
ggplot object (print to show the plot)
}
\description{
Construct a scatterplot comparing the p-value to the quantile of Z,
coloring points either based on an oracle TRUE/FALSE (if the column
\code{oracle} is included) or based on thresholds of the functional
qvalue.
}

