% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.fpi0.R
\name{plot_MISE}
\alias{plot_MISE}
\title{Plot demonstrating how lambda is chosen to minimize MISE}
\usage{
plot_MISE(x, ...)
}
\arguments{
\item{x}{fpi0 object}

\item{...}{Additional arguments (not used)}
}
\description{
Construct a plot containing two panels. The first compares the curve
estimated from each choice of lambda to a reference obtained from
the lowest lambda value (used to estimate how accurate the shape is).
The second panel shows how omega (error of the shape), delta.sq
(squared bias), and their sum (estimate of MISE) change with changing
lambda. The software's choice of lambda is highlighted as a vertical
dashed line.
}

