% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estFPi0.R
\name{predict.fPi0}
\alias{predict.fPi0}
\title{Predict pi0 for a given z value}
\usage{
\method{predict}{fPi0}(object, z0 = NULL, z = NULL, lambda = NULL, ...)
}
\arguments{
\item{object}{fPi0 object}

\item{z0}{new z0 values (before transforming to uniform)}

\item{z}{new z values (after transforming to uniform)}

\item{lambda}{Lambda used for prediction. If null, defaults to the lambda
chosen in the fPi0 object}

\item{...}{Extra arguments, not used}
}
\value{
Vector of fPi0 predictions
}
\description{
Can be given either z0 values (same scale as was given to estFPi0)
or z values (after z0 was transformed to uniform; this is the scale
on which plots of fpi0 are made)
}

