% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_fpi0.R
\name{predict.fPi0}
\alias{predict.fPi0}
\title{Predict the functional proportion}
\usage{
\method{predict}{fPi0}(object, z0 = NULL, z = NULL, lambda = NULL,
  ...)
}
\arguments{
\item{object}{fPi0 object}

\item{z0}{New z0 values (when they are between 0 and 1)}

\item{z}{New z values (as the quantile transformed values of z0)}

\item{lambda}{The value of the tuning parameter to be used for prediction. If null, defaults to the lambda
chosen in the fPi0 object}

\item{...}{Extra arguments, not used}
}
\value{
Vector of fPi0 predictions
}
\description{
Predict the values of the functional proportion when it is evaluated at a vector z' whose entries are between 0 and 1. The vector z' can contain observations directly from the informative variable z0 when z0 is between 0 and 1, or contain the quantile transformed observations of z0 that are regarded as observations from the quantile tranformed variable z. For information on z, please refer to \code{fqvalue} or \code{estimate_fpi0}
}
