% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fqvalue.R
\name{summary.fqvalue}
\alias{summary.fqvalue}
\title{One row summary of an analysis by the fFDR methodology}
\usage{
\method{summary}{fqvalue}(object, oracle, ...)
}
\arguments{
\item{object}{An object of class fqvalue.}

\item{oracle}{A logical value TRUE/FALSE, where TRUE means
the oracle procedure is available, and FALSE the contrary}

\item{...}{Extra arguments, not used}
}
\value{
A tbl_df with columns summarizing the performance of the fFDR methodology in terms of the fqvalue and FDR, with comparison to those obtained by the traditional qvalue method.
}
\description{
One row summary of an analysis by the fFDR methodology
}
