% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.fqvalue.R
\name{summary.fqvalue}
\alias{summary.fqvalue}
\title{One row summary of an fqvalue analysis}
\usage{
\method{summary}{fqvalue}(object, oracle, ...)
}
\arguments{
\item{object}{An object of class fqvalue.}

\item{oracle}{A logical vector giving the oracle, where TRUE means
the alternative hypothesis holds and FALSE means the null does}

\item{...}{Extra arguments, not used}
}
\value{
A tbl_df with columns summarizing the performance of fqvalue
}
\description{
One row summary of an fqvalue analysis
}

